Support multiple AMQP servers
In your config you can specify a list of AMQP servers:

`RABBIT_HOST:foo,bar,baz`

One of the servers from the list will be picked and connected to. If the
connection fails we will keep trying to pick a new server from the list
and connect to it until we are successful.

Config tests updated to reflect new settings.