Use settings to reconnect to the server properly
Pass the AMQP settings object into the reconnect_to method. This will
ensure that the correct details are used when connecting to a new host.

Switch on auto_recovery on the channel so that the queues and exchanges
are setup properly again (See the amqp ErrorHandling doc for info).

Remove rspec/mocks entry which was accidentally checked in.

Update config docs to show new RABBIT_HOST example.